ribaj: Add the ability to query by regions
Riba has an additional concept of "regions" which are like
sections but are only one level deep and only have an article
layout.

This change simply changes the `prepare-auto-query` module
to check for regions and add them to the query if they're
present on the list.

Also need to add an article schema (with regions) so that they
can be added to the article models.